// src/generated/client/models/WorkShiftChangeReason.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The reason why the change entry has been added to the change set.
 * @export
 */
export const WorkShiftChangeReason = {
    WorkshiftCreated: 'WORKSHIFT_CREATED',
    WorkshiftUpdatedApproved: 'WORKSHIFT_UPDATED_APPROVED',
    WorkshiftUpdatedDayoffworkallowance: 'WORKSHIFT_UPDATED_DAYOFFWORKALLOWANCE',
    WorkshiftUpdatedAbsence: 'WORKSHIFT_UPDATED_ABSENCE',
    WorkshiftUpdatedPerdiemallowance: 'WORKSHIFT_UPDATED_PERDIEMALLOWANCE',
    WorkshiftUpdatedNotes: 'WORKSHIFT_UPDATED_NOTES',
    WorkshifthoursUpdatedActualhours: 'WORKSHIFTHOURS_UPDATED_ACTUALHOURS',
    WorkeventUpdatedCostcenter: 'WORKEVENT_UPDATED_COSTCENTER',
    WorkeventUpdatedType: 'WORKEVENT_UPDATED_TYPE',
    WorkeventUpdatedTimestamp: 'WORKEVENT_UPDATED_TIMESTAMP'
} as const;
export type WorkShiftChangeReason = typeof WorkShiftChangeReason[keyof typeof WorkShiftChangeReason];


export function WorkShiftChangeReasonFromJSON(json: any): WorkShiftChangeReason {
    return WorkShiftChangeReasonFromJSONTyped(json, false);
}

export function WorkShiftChangeReasonFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkShiftChangeReason {
    return json as WorkShiftChangeReason;
}

export function WorkShiftChangeReasonToJSON(value?: WorkShiftChangeReason | null): any {
    return value as any;
}