// src/generated/client/models/WorkShiftChangeSet.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Vehicle Management Services (management)
 * Vehicle Management Services (management)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { WorkShiftChange } from './WorkShiftChange';
import {
    WorkShiftChangeFromJSON,
    WorkShiftChangeFromJSONTyped,
    WorkShiftChangeToJSON,
} from './WorkShiftChange';

/**
 * Represents a single set of changes happened withing single work shift.
 * 
 * When a manager creates or updates a work shift or its work shift hours, a client-generated WORKSHIFT_CHANGE_ID
 * will always be sent with the API requests. The ID is used to link the changes to correct change set. The change
 * set contains a list of changes that describe the changes made to the work shift.
 * 
 * @export
 * @interface WorkShiftChangeSet
 */
export interface WorkShiftChangeSet {
    /**
     * 
     * @type {string}
     * @memberof WorkShiftChangeSet
     */
    readonly id?: string;
    /**
     * Creation time
     * @type {Date}
     * @memberof WorkShiftChangeSet
     */
    readonly createdAt?: Date;
    /**
     * Creator's ID
     * @type {string}
     * @memberof WorkShiftChangeSet
     */
    readonly creatorId?: string;
    /**
     * 
     * @type {Array<WorkShiftChange>}
     * @memberof WorkShiftChangeSet
     */
    readonly entries?: Array<WorkShiftChange>;
}

/**
 * Check if a given object implements the WorkShiftChangeSet interface.
 */
export function instanceOfWorkShiftChangeSet(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function WorkShiftChangeSetFromJSON(json: any): WorkShiftChangeSet {
    return WorkShiftChangeSetFromJSONTyped(json, false);
}

export function WorkShiftChangeSetFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkShiftChangeSet {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'createdAt': !exists(json, 'createdAt') ? undefined : (new Date(json['createdAt'])),
        'creatorId': !exists(json, 'creatorId') ? undefined : json['creatorId'],
        'entries': !exists(json, 'entries') ? undefined : ((json['entries'] as Array<any>).map(WorkShiftChangeFromJSON)),
    };
}

export function WorkShiftChangeSetToJSON(value?: WorkShiftChangeSet | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
    };
}